week pagination with pagy OK
