+++ app/controllers/events_controller.rb
@@ -3,6 +3,7 @@ class EventsController < ApplicationController
   before_action :set_events, only: %i[index map calendar]
 
   include Pagy::Backend
+  include PagyCalendar
 
   def index
   end
@@ -35,26 +36,10 @@ class EventsController < ApplicationController
 
   private
 
-  # REQUIRED: return the start and end limits of the collection as a 2 items array
-  def pagy_calendar_period(collection)
-    starting = collection.minimum(:date)
-    ending = collection.maximum(:date)
-    [ starting.in_time_zone, ending.in_time_zone ]
-  end
-
-  # REQUIRED: return the collection filtered by a time period
-  def pagy_calendar_filter(collection, from, to)
-    collection.where(date: from...to)
-  end
-
-  # OPTIONAL: return the array counts per time
-  def pagy_calendar_counts(collection, unit, from, to)
-    collection.group_by_period(unit, :date, range: from...to).count.values
-  end
-
   def set_events
     events = authorize policy_scope(Event)
-    @calendar, @pagy, @events = pagy_calendar(events, day: {}, pagy: {})
+    @calendar, @pagy, @events = pagy_calendar(events, week: {}, pagy: {limit: false})
+    @next_week_page = @calendar[:week].next
     set_events_days
   end
